+++ src/main/java/vanilla/java/chronicle/tcp/InProcessChronicleSource.java
@@ -103,7 +103,7 @@ public class InProcessChronicleSource<C extends Chronicle> implements Chronicle 
 
                     bb.clear();
                     bb.putLong(index);
-                    bb.putLong(size);
+                    bb.putInt(size);
                     while (remaining > 0) {
                         int size2 = Math.min(remaining, bb.capacity());
                         bb.limit(size2);
